requirements.txt
Django==1.7
GitPython==0.3.2.1
MySQL-python==1.2.5
Pillow==2.6.1
argparse==1.2.1
django-appconf==0.6
django-ckeditor==4.4.6
django-contact-form==1.0
-e git+https://github.com/nunataksoftware/django-deployer.git@9d7b720c1aa0e39a17cc9fd4f1bf9bda48ef4d0b#egg=django_deployer-master
django-imagekit==3.2.4
gitdb==0.6.0
pilkit==1.1.12
six==1.8.0
smmap==0.8.3
wsgiref==0.1.2
